requirements.txt
tqdm~=4.64.1
numpy~=1.23.1
matplotlib~=3.6.0
torchaudio~=0.13.0